# server/swaggerDoc.json
{
	"swagger": "2.0",
	"info": {
		"description": "Demo",
		"version": "1.0",
		"title": "Nada-Demo API文档"
	},
	"host": "10.100.83.171:8010",
	"basePath": "/",
...
	"paths": {
...
		"/corp/get/{id}": {
			"get": {
				"tags": [
					"公司操作接口"
				],
				"summary": "获取公司详细信息(id)",
				"description": "根据公司id获取匹配的公司信息",
				"operationId": "getUsingGET_1",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "公司id",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/AResultSet«Corp»"
						}
					}
				}
			}
		},
...
		"/user/get/{id}": {
			"get": {
				"tags": [
					"用户操作接口"
				],
				"summary": "获取用户详细信息(id)",
				"description": "根据用户id获取匹配的用户信息",
				"operationId": "getUsingGET_2",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "公司id",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/AResultSet«User»"
						}
					}
				}
			}
		},
...
	},
	"securityDefinitions": {
		"spring_oauth": {
			"type": "oauth2",
			"authorizationUrl": "null/authorize",
			"tokenUrl": "null/token",
			"flow": "accessCode",
			"scopes": {
				"read": "for read operations",
				"write": "for write operations",
				"foo": "Access foo API"
			}
		}
	},
	"definitions": {
...
		"AData«Corp»": {
			"type": "object",
			"properties": {
				"count": {
					"type": "integer",
					"format": "int64",
					"description": "数据总数",
					"allowEmptyValue": false
				},
				"list": {
					"description": "数据",
					"allowEmptyValue": false,
					"$ref": "#/definitions/Corp"
				},
				"pageNo": {
					"type": "integer",
					"format": "int32",
					"description": "当前页",
					"allowEmptyValue": false
				},
				"pageSize": {
					"type": "integer",
					"format": "int32",
					"description": "页数据数",
					"allowEmptyValue": false
				}
			},
			"title": "AData«Corp»"
		},
...
		"AData«User»": {
			"type": "object",
			"properties": {
				"count": {
					"type": "integer",
					"format": "int64",
					"description": "数据总数",
					"allowEmptyValue": false
				},
				"list": {
					"description": "数据",
					"allowEmptyValue": false,
					"$ref": "#/definitions/User"
				},
				"pageNo": {
					"type": "integer",
					"format": "int32",
					"description": "当前页",
					"allowEmptyValue": false
				},
				"pageSize": {
					"type": "integer",
					"format": "int32",
					"description": "页数据数",
					"allowEmptyValue": false
				}
			},
			"title": "AData«User»"
		},
...
		"AResultSet«Corp»": {
			"type": "object",
			"properties": {
				"data": {
					"$ref": "#/definitions/AData«Corp»"
				},
				"message": {
					"type": "string",
					"description": "信息",
					"allowEmptyValue": false
				},
				"status": {
					"type": "integer",
					"format": "int32",
					"description": "状态编码",
					"allowEmptyValue": false
				}
			},
			"title": "AResultSet«Corp»",
			"description": "统一返回结果集类（非实体类）"
		},
...
		"AResultSet«User»": {
			"type": "object",
			"properties": {
				"data": {
					"$ref": "#/definitions/AData«User»"
				},
				"message": {
					"type": "string",
					"description": "信息",
					"allowEmptyValue": false
				},
				"status": {
					"type": "integer",
					"format": "int32",
					"description": "状态编码",
					"allowEmptyValue": false
				}
			},
			"title": "AResultSet«User»",
			"description": "统一返回结果集类（非实体类）"
		},
...
		"Corp": {
			"type": "object",
			"properties": {
				"id": {
					"type": "string",
					"example": {
						"combo": 1
					},
					"description": "公司ID",
					"allowEmptyValue": false,
					"minLength": 0,
					"maxLength": 5
				},
				"corpName": {
					"type": "string",
					"description": "公司名",
					"allowEmptyValue": false,
					"minLength": 0,
					"maxLength": 50
				},
				"managerID": {
					"type": "string",
					"example": {
						"combo": 3
					},
					"description": "管理员ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"parentCorpID": {
					"type": "string",
					"example": {
						"combo": 1
					},
					"description": "上级公司ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"corpEmpTotals": {
					"type": "integer",
					"format": "int32",
					"description": "总员工数",
					"allowEmptyValue": true,
					"minimum": 0,
					"maximum": 10,
					"exclusiveMinimum": false,
					"exclusiveMaximum": false
				},
				"useYN": {
					"type": "integer",
					"format": "int32",
					"example": {
						"combo": 4
					},
					"description": "使用与否",
					"allowEmptyValue": true,
					"minimum": 0,
					"maximum": 2,
					"exclusiveMinimum": false,
					"exclusiveMaximum": false
				},
				"createUser": {
					"type": "string",
					"description": "创建用户ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"createDate": {
					"type": "string",
					"format": "date-time",
					"description": "创建日期",
					"allowEmptyValue": true
				},
				"updateUser": {
					"type": "string",
					"description": "修改用户ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"updateDate": {
					"type": "string",
					"format": "date-time",
					"description": "修改日期",
					"allowEmptyValue": true
				}
			},
			"title": "Corp",
			"description": "公司"
		},
...
		"User": {
			"type": "object",
			"properties": {
				"id": {
					"type": "string",
					"description": "用户ID",
					"allowEmptyValue": false,
					"minLength": 0,
					"maxLength": 10
				},
				"corpID": {
					"type": "string",
					"example": {
						"combo": 1
					},
					"description": "公司ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 5
				},
				"userName": {
					"type": "string",
					"description": "用户名",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 50
				},
				"mobileNo": {
					"type": "string",
					"description": "手机号",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 20
				},
				"telNo": {
					"type": "string",
					"description": "座机号",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 20
				},
				"email": {
					"type": "string",
					"description": "邮箱",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 50
				},
				"useYN": {
					"type": "integer",
					"format": "int32",
					"example": {
						"combo": 4
					},
					"description": "使用与否",
					"allowEmptyValue": true,
					"minimum": 0,
					"maximum": 2,
					"exclusiveMinimum": false,
					"exclusiveMaximum": false
				},
				"posID": {
					"type": "string",
					"description": "职位ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"dutyID": {
					"type": "string",
					"description": "职责ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"password": {
					"type": "string",
					"description": "密码",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 100
				},
				"deptGroup": {
					"type": "string",
					"example": {
						"combo": 2
					},
					"description": "部门组织ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"createUser": {
					"type": "string",
					"description": "创建用户ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"createDate": {
					"type": "string",
					"format": "date-time",
					"description": "创建日期",
					"allowEmptyValue": true
				},
				"updateUser": {
					"type": "string",
					"description": "修改用户ID",
					"allowEmptyValue": true,
					"minLength": 0,
					"maxLength": 10
				},
				"updateDate": {
					"type": "string",
					"format": "date-time",
					"description": "修改日期",
					"allowEmptyValue": true
				}
			},
			"title": "User",
			"description": "用户"
		},
...
	}
}